+++ libnavigation-core/src/main/java/com/mapbox/navigation/core/trip/service/NavigationNotificationService.kt
@@ -5,8 +5,13 @@ import android.app.Service
 import android.content.Context
 import android.content.Intent
 import android.os.IBinder
+import androidx.annotation.CallSuper
 import androidx.core.app.ServiceCompat
+import com.mapbox.navigation.core.internal.dump.MapboxDumpHandler
+import com.mapbox.navigation.core.internal.dump.MapboxDumpRegistry
 import com.mapbox.navigation.core.telemetry.MapboxNavigationTelemetry
+import java.io.FileDescriptor
+import java.io.PrintWriter
 
 /**
  * Service is updating information about current trip
@@ -17,6 +22,13 @@ internal class NavigationNotificationService : Service() {
         notificationResponse.notification.flags = Notification.FLAG_FOREGROUND_SERVICE
         startForeground(notificationResponse.notificationId, notificationResponse.notification)
     }
+
+    /**
+     * This will handle commands from `adb shell dumpsys activity service`.
+     *
+     * Use the [MapboxDumpRegistry] to add or remove dump interceptors.
+     */
+    private val mapboxDumpHandler = MapboxDumpHandler()
 
     /**
      * Return the communication channel to the service (always *null*)
@@ -46,5 +58,13 @@ internal class NavigationNotificationService : Service() {
         super.onDestroy()
         ServiceCompat.stopForeground(this, ServiceCompat.STOP_FOREGROUND_REMOVE)
         MapboxTripService.unregisterOneTimeNotificationDataObserver(notificationDataObserver)
+    }
+
+    /**
+     * Overrides the dump command so that state can be changed with adb.
+     */
+    @CallSuper
+    override fun dump(fd: FileDescriptor, writer: PrintWriter, args: Array<String>?) {
+        mapboxDumpHandler.handle(fd, writer, args)
     }
 }